make held_out set optional

+++ corebehrt/main/create_data.py
@@ -82,10 +82,11 @@ def main_data(config_path):
     logger.info("Tokenizing tuning")
     load_tokenize_and_save(features_path, tokenizer, tokenized_path, "tuning")
 
-    logger.info("Tokenizing held_out")
-    load_tokenize_and_save(features_path, tokenizer, tokenized_path, "held_out")
-    torch.save(tokenizer.vocabulary, join(tokenized_path, "vocabulary.pt"))
-    logger.info("Finished tokenizing")
+    if os.path.exists(os.path.join(features_path, "held_out")):
+        logger.info("Tokenizing held_out")
+        load_tokenize_and_save(features_path, tokenizer, tokenized_path, "held_out")
+        torch.save(tokenizer.vocabulary, join(tokenized_path, "vocabulary.pt"))
+        logger.info("Finished tokenizing")
 
 
 if __name__ == "__main__":
